rainbow: Implement board detector
It seems to work well enough at picking out the sides. Works better with
higher resolutions.

Finding the bottom of the board is dependent on knowing the color of the
board, so that we can correctly ignore other horizontal lines. That
might be worth pulling in to the generic horizon detector.